--- Assets/PlayerAim.cs
@@ -1,78 +1,0 @@
-using UnityEngine;
-
-    public class PlayerAim : MonoBehaviour
-    {
-        #region Datamembers
-
-        #region Editor Settings
-
-        [SerializeField] private LayerMask groundMask;
-
-        #endregion
-        #region Private Fields
-
-        private Camera mainCamera;
-        private Vector3 Direction;
-
-        #endregion
-
-        #endregion
-
-
-        #region Methods
-
-        #region Unity Callbacks
-
-        private void Start()
-        {
-            // Cache the camera, Camera.main is an expensive operation.
-            mainCamera = Camera.main;
-        }
-
-        private void Update()
-        {
-            Aim();
-        }
-
-        #endregion
-
-        private void Aim()
-        {
-            var (success, position) = GetMousePosition();
-            if (success)
-            {
-                // Calculate the direction
-                Direction = position - transform.position;
-
-                // You might want to delete this line.
-                // Ignore the height difference.
-                Direction.y = 0;
-
-                // Make the transform look in the direction.
-                transform.forward = Direction;
-            }
-        }
-
-        private (bool success, Vector3 position) GetMousePosition()
-        {
-            var ray = mainCamera.ScreenPointToRay(Input.mousePosition);
-
-            if (Physics.Raycast(ray, out var hitInfo, Mathf.Infinity, groundMask))
-            {
-                // The Raycast hit something, return with the position.
-                return (success: true, position: hitInfo.point);
-            }
-            else
-            {
-                // The Raycast did not hit anything.
-                return (success: false, position: Vector3.zero);
-            }
-        }
-
-    public Vector3 GetDirection()
-    {
-        return this.Direction;
-    }
-
-        #endregion
-    }
