+++ Assets/Scripts/PlayerAim.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5599fa7e0cb228e4bbf1a2f7d0d1fb39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
